requirements.txt
# AutoBasket Betting Intelligence
# Зависимости проекта

# === CORE ===
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# === MACHINE LEARNING ===
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
anthropic>=0.18.0

# === WEB & API ===
requests>=2.31.0
aiohttp>=3.8.0
asyncio>=3.4.3
httpx>=0.24.0

# === NBA DATA ===
nba_api>=1.4.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# === DATABASE ===
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# === DASHBOARD ===
streamlit>=1.28.0
plotly>=5.17.0

# === NOTIFICATIONS ===
python-telegram-bot>=20.0
discord-webhook>=1.3.0

# === UTILITIES ===
python-dotenv>=1.0.0
pydantic>=2.0.0
loguru>=0.7.0
schedule>=1.2.0

# === TESTING ===
pytest>=7.4.0
pytest-asyncio>=0.21.0
